Introduced AbstractTransaction
