+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/AbstractTransaction.java
@@ -1,0 +1,24 @@
+package com.concordium.sdk.transactions;
+
+import lombok.Setter;
+
+public class AbstractTransaction implements Transaction {
+
+    @Setter
+    private BlockItem item;
+
+    @Override
+    public byte[] getBytes() {
+        return item.getVersionedBytes();
+    }
+
+    @Override
+    public Hash getHash() {
+        return item.getHash();
+    }
+
+    @Override
+    public int getNetworkId() {
+        return DEFAULT_NETWORK_ID;
+    }
+}
